+++ source/Jitter/Collision/Shapes/Shape.cs
@@ -47,7 +47,7 @@ namespace Jitter.Collision.Shapes
 
         // internal values so we can access them fast  without calling properties.
         internal JMatrix inertia = JMatrix.Identity;
-        internal float mass = 1.0f;
+        internal double mass = 1.0f;
 
         internal JBBox boundingBox = JBBox.LargeBox;
         internal JVector geomCen = JVector.Zero;
@@ -73,7 +73,7 @@ namespace Jitter.Collision.Shapes
         /// <summary>
         /// Gets the mass of the shape. This is the volume. (density = 1)
         /// </summary>
-        public float Mass { get { return mass; } protected set { mass = value; } }
+        public double Mass { get { return mass; } protected set { mass = value; } }
 
         /// <summary>
         /// Informs all listener that the shape changed.
@@ -111,7 +111,7 @@ namespace Jitter.Collision.Shapes
         /// <param name="generationThreshold"></param>
         public virtual void MakeHull(ref List<JVector> triangleList, int generationThreshold)
         {
-            float distanceThreshold = 0.0f;
+            double distanceThreshold = 0.0f;
 
             if (generationThreshold < 0) generationThreshold = 4;
 
@@ -163,9 +163,9 @@ namespace Jitter.Collision.Shapes
                 JVector p2; SupportMapping(ref tri.n2, out p2);
                 JVector p3; SupportMapping(ref tri.n3, out p3);
 
-                float d1 = (p2 - p1).LengthSquared();
-                float d2 = (p3 - p2).LengthSquared();
-                float d3 = (p1 - p3).LengthSquared();
+                double d1 = (p2 - p1).LengthSquared();
+                double d2 = (p3 - p2).LengthSquared();
+                double d3 = (p1 - p3).LengthSquared();
 
                 if (Math.Max(Math.Max(d1, d2), d3) > distanceThreshold && tri.generation < generationThreshold)
                 {
@@ -281,11 +281,11 @@ namespace Jitter.Collision.Shapes
         /// <param name="centerOfMass"></param>
         /// <param name="inertia">Returns the inertia relative to the center of mass, not to the origin</param>
         /// <returns></returns>
-        #region  public static float CalculateMassInertia(Shape shape, out JVector centerOfMass, out JMatrix inertia)
-        public static float CalculateMassInertia(Shape shape, out JVector centerOfMass,
+        #region  public static double CalculateMassInertia(Shape shape, out JVector centerOfMass, out JMatrix inertia)
+        public static double CalculateMassInertia(Shape shape, out JVector centerOfMass,
             out JMatrix inertia)
         {
-            float mass = 0.0f;
+            double mass = 0.0f;
             centerOfMass = JVector.Zero; inertia = JMatrix.Zero;
 
             if (shape is Multishape) throw new ArgumentException("Can't calculate inertia of multishapes.", "shape");
@@ -296,7 +296,7 @@ namespace Jitter.Collision.Shapes
 
             // create inertia of tetrahedron with vertices at
             // (0,0,0) (1,0,0) (0,1,0) (0,0,1)
-            float a = 1.0f / 60.0f, b = 1.0f / 120.0f;
+            double a = 1.0f / 60.0f, b = 1.0f / 120.0f;
             JMatrix C = new JMatrix(a, b, b, b, a, b, b, b, a);
 
             for (int i = 0; i < hullTriangles.Count; i += 3)
@@ -309,14 +309,14 @@ namespace Jitter.Collision.Shapes
                     column0.Y, column1.Y, column2.Y,
                     column0.Z, column1.Z, column2.Z);
 
-                float detA = A.Determinant();
+                double detA = A.Determinant();
 
                 // now transform this canonical tetrahedron to the target tetrahedron
                 // inertia by a linear transformation A
                 JMatrix tetrahedronInertia = JMatrix.Multiply(A * C * JMatrix.Transpose(A), detA);
 
                 JVector tetrahedronCOM = (1.0f / 4.0f) * (hullTriangles[i + 0] + hullTriangles[i + 1] + hullTriangles[i + 2]);
-                float tetrahedronMass = (1.0f / 6.0f) * detA;
+                double tetrahedronMass = (1.0f / 6.0f) * detA;
 
                 inertia += tetrahedronInertia;
                 centerOfMass += tetrahedronMass * tetrahedronCOM;
@@ -326,9 +326,9 @@ namespace Jitter.Collision.Shapes
             inertia = JMatrix.Multiply(JMatrix.Identity, inertia.Trace()) - inertia;
             centerOfMass = centerOfMass * (1.0f / mass);
 
-            float x = centerOfMass.X;
-            float y = centerOfMass.Y;
-            float z = centerOfMass.Z;
+            double x = centerOfMass.X;
+            double y = centerOfMass.Y;
+            double z = centerOfMass.Z;
 
             // now translate the inertia by the center of mass
             JMatrix t = new JMatrix(
