+++ source/Jitter/Dynamics/Constraint.cs
@@ -32,7 +32,7 @@ namespace Jitter.Dynamics.Constraints
 
     public interface IConstraint
     {
-        void PrepareForIteration(float timestep);
+        void PrepareForIteration(double timestep);
         void Iterate();
 
         RigidBody Body1 { get; }
@@ -88,7 +88,7 @@ namespace Jitter.Dynamics.Constraints
         /// Called once before iteration starts.
         /// </summary>
         /// <param name="timestep">The simulation timestep</param>
-        public abstract void PrepareForIteration(float timestep);
+        public abstract void PrepareForIteration(double timestep);
 
         /// <summary>
         /// Iteratively solve this constraint.
